+++ src/light.cpp
@@ -831,7 +831,7 @@ void Level::lights_render_small_lights(int minx, int miny, int maxx, int maxy, i
             continue;
           }
 
-          auto mid = (blit_br + blit_tl) / 2;
+          auto mid = (blit_br + blit_tl) / (short) 2;
 
           float light_scale = l->light_dist;
 
@@ -922,7 +922,7 @@ void Level::lights_render_small_lights(int minx, int miny, int maxx, int maxy, i
           }
 
           auto  s   = l->light_dist + l->flicker;
-          auto  mid = (blit_br + blit_tl) / 2;
+          auto  mid = (blit_br + blit_tl) / (short) 2;
           auto  tlx = mid.x - s;
           auto  tly = mid.y - s;
           auto  brx = mid.x + s;
